Added the DM private "pydoc" command
This private DM message responds to user requests with the results of
pydoc help output. The command recognizes a -k switch for keyword
searches as well as specific keywords: help, modules topics, and the
specific capitialized topic names listed by topics.

A help message is returned if the command does not recognize a switch,
gets zero length text, the text is 'help' or the pydoc help()
invocation indicates it failed to find documentation for the submitted
text.